enhance function to process data input
updated
